+++ LiDARC-backend/src/main/java/com/example/lidarcbackend/service/comparisons/ComparisonService.java
@@ -250,7 +250,7 @@ public class ComparisonService implements IComparisonService {
     }
 
     private ComparisonPlan processFolderGroup(List<Long> fileIds, GridParameters grid, Long comparisonId,
-            String groupName, Integer pointFilterLowerBound, Integer pointFilterUpperBound,
+            String groupName, Double pointFilterLowerBound, Double pointFilterUpperBound,
             Boolean outlierDetectionEnabled, Double outlierDeviationFactor, Boolean needPointFilter)
             throws NotFoundException {
         ComparisonPlan plan = new ComparisonPlan();
